requirements.txt
# Core LangChain dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-chroma>=0.1.0
langchain-community>=0.0.20
langchainhub>=0.1.14

# Vector database
chromadb>=0.4.22

# Embeddings
sentence-transformers>=2.2.2
openai>=1.12.0

# PDF processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# Text processing
tiktoken>=0.5.2

# Web framework
streamlit>=1.31.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Utilities
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

